Update to the text-interaction mode for SMT solvers.
This update fixes a variety of issues with solver interaction involving
the named assertion and unsat core updates.  CVC4, Z3 and Yices now will
properly issue queries and return resuts.

In addition, a variety of refactorings were made to the interaction code
that should make things more uniform going forward.  As part of these
refactorings, it is now possible to pass an "auxiliary output" file handle
when interacting with solvers.  If provided, solver commands and responses
will be echoed into this file for debugging purposes.