Add initial FP80 representation to What4.
This is the minimum needed for allowing the presence of FP80 (80-bit
floating point values) in equations.  It only supports them in
IEEE754-style representations, and even there the semantics may not be
correct, so caution should be taken if there is any examination of or
reliance on these values at this point.