Fix(?) online solver process termination
The `interruptProcessGroupOf` function from System.Process sends SIGINT, which
is the equivalent of Ctrl+C.  I'm not sure if Ctrl+C behavior is quite the right
notion we want to capture; I think SIGTERM (signal 15 on x86/Linux and most
other platforms) is the "soft kill" that is still handleable.

Selfishly, I want to make this change because yices (in native non-SMT mode) has
special handling for SIGINT that doesn't work when MCSAT is enabled.  yices-smt2
has the correct behavior.

Switching to SIGTERM (via `terminateProcess`) has the expected behavior on
yices, z3, and cvc4.

Note that this is not the uncatchable SIGKILL (usually signal 9).