what4: Prevent handle errors in the dReal adapter
The removed code copies stdout to the log handle.  This seems to cause handle
errors like:

```
fd:5: hGetBufSome: illegal operation (handle is closed)
```

I assume that is because the forked thread outlives the scope of the process
connection (inside of `withProcessHandles`), which closes handles when it
completes.  The boolector backend, which seems fairly similar in structure, does
not have this removed code.