Develop a precise test for mixed-domain overlap.
This turns out to be pretty tricky, but the Cryptol functions in this
patch demonstrate that we can compute a precise overlap test for mixed
bitvector domains without resorting to brute enumeration or other
inefficent methods.

The main operation of interest is the `bitwise_round_between` method,
which finds the arithmetically next value above some starting value
that is bitwise between consistent bitbounds.  This turns out to
be precisely what we need to test whether an arithmetic interval
which is nested "inside" a bitwise interval actually contains one
of the bitwise values.

It might be possible to simplify this operation, but the current
definition demonstrates it is possible using only word-level operations,
if we allow ourselves to use the `fillright` operation.