Allow clients to select what type of program locations they want to use.
In order to make a mimimally invasive change to the API, the `t`
parameter to `ExprBuilder` and related types is repurposed via
type families to carry both the NonceGenerator type brand and to
select the type of program locations.  This requires slightly more
setup when setting up an `ExprBuilder`, but keeps a fairly compatible
API.

As a result, the `ProgramLoc` and `FunctionName` modules are
no longer used inside What4 and will be moved into `crucible`,
where they are a better fit.