Add alternate code paths for online solvers that do not support `reset-assertions` or `get-info` commands.
We add a class method `getErrorBehavior` which defaults to
returning `ImmediateExit`, and actually queries the solver
itself only for solvers that implement the appropriate `get-info`
command.

For solvers that do not support `reset-assertions`, we instead
simulate the desired behavior by immediately pushing an assertion
frame on startup and popping all outstanding frames when
a reset is requested.