Fix incorrect range reduction code in SWord shifts.
When using the SWord module for dynamically-sized words, one has to
match the sizes of shift operations before passing them down to the
main representation code. When the shift amount has a larger bit-width
than the word to be shifted, a range reduction is required so the
truncation of the shift amount still gets the correct result. Previously,
this reduction was being done incorrectly by computing a `max` instead
of a `min`.

Fixes #126