Add resolveSymBV
Sometimes, `{un,}signedBVBounds` aren't sufficient to determine the lower and
upper bounds of a `SymBV`. For these situations, it can be helpful to query the
SMT solver to search for the bounds, later using `unsafeSetAbstractValue` to
update the newly discovered bounds to unlock further What4 simplifications.
The `resolveSymBV` function does exactly this kind of search. I've added
`hedgehog` tests which ensure that it behaves the way you'd expect it to.