template_tests: Emit an error code if tests fail
Running `hedgehog`'s `checkSequential` function alone isn't enough to cause a
test suite to fail if one or more of the test cases to fail. Per a comment in
the `hedgehog` source code
[here](https://github.com/hedgehogqa/haskell-hedgehog/blob/3d4dde46e15ab25a10cac0deff03bae10ecacc4e/hedgehog-example/test/test.hs#L34-L42),
the idiomatic way to do make a `checkSequential` failure imply test suite
failure is to check the result of running `checkSequential` and, if that is
`False`, throw `exitFailure`. This patch incorporates that advice into the
`test_templates` test suite.

Fixes #206.