Merge what4-serialize into what4 core (#217)
This change incorporates the serialization/deserialization capabilities from the what4-serialize package into what4. The new functionality enables serializing what4 terms as s-expressions that are intended to match SMTLib2. It also supports deserializing them from the text form into what4 terms in memory.

Note that this is just an import of the code from the what4-serialize library (with a few modest fixes), so if you previously relied on what4-serialize, you will need to delete your dependency on that library when updating. Post-merge, the code in what4 should be a drop-in replacement for the what4-serialize library.

Fixes #104 

Co-authored-by: Ryan Scott <ryan.gl.scott@gmail.com>