WIP: Equalities
This commit introduces a new data structure `Equalities` for storing
conjunctions of (dis)equalities, based on an (annotated) union-find.
Much like `ConjMap` for `Bool`, this data structure implements several
algebraic rules "for free", notably symmetry and transitivity. The
hope is that What4 can keep more expressions concrete by performing
deductions via (1) symmetry and transitivity and (2) resolution of
conflicting equalities and disequalities.

While this appears to make `BaseEq` more complex, possibly raising
concerns about memory usage, it is worth recalling that a union find
can store `n^2` equalities in `n` space. Hopefully, this compression
outweighs the additional complexity of the internal maps in applications
we care about.

It is possible to represent conjunctions of possibly negated terms
using equality, e.g., `a /\ ~b <==> a = True /\ b = False`. Representing
conjunctions this way in `Equalities` would have all of the same
normalization properties as `BoolMap` (specifically, associativity,
commutativity, and idempotency). It would likely be a win to move in
this direction.