ci: Move `cabal check` to the end
Changing the order in which the CI system performs certain steps can
cause builds that are going to fail to do so faster. If all steps are
successful, then the relative order of independent steps (steps that
don’t depend on one another) is irrelevant: it won’t change the overall
build time. However, if one of the steps is going to fail, then the
earlier that step occurs in the CI build, the faster the build will
fail. This results in faster feedback for the developer. All other
things being equal, steps that fail more often should occur earlier in
the CI workflow.

Our Cabal files rarely change, so cabal check almost always passes, so
it's safe to put it at the end to get faster feedback from the build and
test stages.