Simplify implementations of {bv,sbv}IntTerm
The code in `sbvIntTerm` was way more complicated than it needed to be.
Moreover, it was doing basically the same thing as `bvIntTerm`, but adding an
additional summand.

This commit simplifies the code in `sbvIntTerm` to avoid unnecessary
argument-passing, and it refactors the code in `sbvIntTerm` and `bvIntTerm` to
make the similarities between the two functions more obvious at a glance.