Fix off-by-one error in sbvIntTerm
This fixes a regression that I inadvertently introduced in `what4-1.7.1` as
part of commit f48d0fa291c73b21510c16cb5866a445834d62a3. The issue is that
`sbvIntTerm w x` needs to compute exactly `w` summands, `w - 1` of which are
the digits computed from the least significant digits, and the last summand
should be the offset in order to make the integer properly signed. The
aforementioned commit contained an off-by-one error that caused the `w` least
significant bits (i.e., all the bits) to be used, not the `w - 1` least
significant bits!

This fixes the error and adds another regression test. I have also improved the
comments in `sbvIntTerm` and the closely related function `bvIntTerm`, on which
`sbvIntTerm`'s implementation was inspired. While the implementations of the
two functions are very close, they do differ in certain key aspects, so I have
drawn closer attention to these differences in the comments.

Fixes #329.